Add the possibility to order the albums by release date in the library.
Fixes #114.

This commit allows the user to select between "alphabet" and "release" in the preferences for the default ordering for albums in the LibraryFragment.
The default setting is "alphabet".

If the albums are sorted by release, the release year is added as prefix in front of the album title, like: "2015 - Album title".
In case of alphabetical ordering just the album title is displayed.

If an artist is added to Clementine's playlist, this setting is respected (therefore the titles are added by release year or alphabet).